Implement Spall-based profiling system with conditional compilation
