+++ example/profiling_demo/main.cpp
@@ -1,4 +1,4 @@
-#include "../../src/profiling.h"
+#include "../../ah.h"
 #include <iostream>
 #include <thread>
 #include <chrono>
