--- src/profiling.cpp
@@ -1,5 +1,0 @@
-#include "profiling.h"
-
-namespace profiling {
-    Profiler g_profiler;
-}
